move validate options from ServiceManagerOptions to ServiceManagerContext
